몬스터 유저를 쫒아가게 설정

+++ CommonLib/CommonLib/data_reader/data_reader_manager.h
@@ -133,9 +133,9 @@ public:
 	};
 	struct Collision
 	{
-		int32_t width;
-		int32_t height;
-		int32_t depth;
+		float width;
+		float height;
+		float depth;
 	};
 	struct MonsterInfo
 	{
